[6.9] "advise against using" is too strong
A specification has many uses, including as a historical reference
document.  W3C might not be "advising against" [any] use, e.g. in the
case of a spec that is "not apparently likely to be adopted"; we may
just be "advising that this specification has not seen implementation
and is therefore obsolete."