Clarify superseding
The process is only meant to be invoked when the REC is not replaced in
place by the obvious next update. In general, just republishing the same
document as REC again is sufficient, as has always been.

This superseding process is meant to help us tie loose ends, not to
generate additional busy work.

Closes #183