Announcements are handled by the Team
Announcements which are required by the process and do not involve a
judgement call or a decision are handled by the Team.
The Team, as per the Process, claims to do so in the name of the
Director, but given that these do not involve a Director's decision,
this claim isn't particularly meaningful.

This commit replaces “the Director” by “the Team” in such cases,
to better reflect reality.

Besides, the Director is defined to be part of the Team anyway, so he
can still make the announcements himself or have an announcement sent on
his behalf in cases where that would feel more appropriate.

This commit does not change any of the Director’s powers.