Expand in place the definition of "minor changes" to existing charters
Prior to this change, the Process allowed for minor changes
to be made to existing charters without an AC Review,
without defining or characterizing what kind of changes count as minor.
Rather than defining a generic notion of "minor changes",
at the risk of having it be used in other parts of the Process
in areas which might call for different criteria,
this commit replaces this use of the term "minor changes" with a
characterization of what types of changes are allowed,
including the most common examples, highlighting the intended use of
this rule.