Grammar fix
In Transition to Proposed Recommendation. a new bullet is introduced,
but its phrasing does not match the subject of all the other bullets in
the list which are in relation to a Working Group or the W3C. The bullet
starting "must not have any substantive changes since…" thus reads that
it is the Working Group or the W3C that must not have any substantive
changes…
This bullet is rephrased to clarify that it is referring to the Working
Group or W3C's Candidate Recommendation Snapshot, and not "substantive
changes" to the working group itself.

Closes #440