Disentangle denying progress and forcing a lower maturity
This is a bit of refactoring, to simplify further edits.
The Team/Director has the ability of the Team to require rec track
documents to be republished at a lower maturity level, but that's only
mentioned in passing in what is otherwise a largely editorial summary of
things better defined elsewhere. Move it to its own section, and
rephrase slightly to give more context as to how that may happen.