editorial: rm undefined COO from informal Note
remove "COO" from a list in an informal Note. "COO" is not defined in the document either (in contrast, CEO is in the glossary), so this seems like it was a mistake that made it past review, and thus should be removed from a non-normative Note. if there's an actual use-case for adding "COO" to the Process, that should be raised as a separate issue for proper definition and consideration.