Adjust the status of Tim Berners Lee in the TAG
This ensures he has permanent access to participate in the TAG,
without putting him on the critical path for any vote or quorum that
involves TAG members, notably the Council's Unanimous Short Circuit.

More broadly, this also removes him from being part of the Council,
as the Council's purpose is to replace his former role as Director in
resolving Formal Objections.

See #784