Limit the involvement of Tim Berners Lee to the TAG proper
As the Council's purpose is to replace his former role as Director in
resolving Formal Objections, this removes him from being part of the
Council. This way, he does not take part in the Council's deliberations,
and is not on the critical path for Council votes or quorum questions,
notably the Council's Unanimous Short Circuit.

See #784