Clarify definition of the REC-track
This makes it clear that a document that uses the various publication
types and transitions of the REC-track is indeed on the REC-track, even
if it isn't planning to go all the way to the end.

Some groups for instance have a declared intention of only going as far
as CR. That doesn't make the documents on the "CR track", as there's no
such thing.

See https://github.com/w3c/strategy/issues/438#issuecomment-1997685426