Clarify confidentiality requirements
This is an attempt to clarify the confidentiality requirements by
separating concerns.

* the second bullet in the list is about using reasonable effort. It's
  not specifically the different levels of confidentiality, nor about any
  particular action that must be taken. It's only about defining the
  standard of care that is applicable to maintaining confidentiality.
* The third bullet is specifically about "must not disclose".

But the phrasing of the second bullet could be read as if there was some
particular task or activity ("effort") to be performed when dealing with
different levels, suggesting that maybe changing levels was OK, as long
as you do it the right way (maybe by redacting something).

This rephrasing clarifies that "reasonable effort" is a general
requirement about confidentiality, and that disclosing beyond the proper
level is not appropriate.