Remove the Proposed Recommendation Phase
Proposed Recommendation is only a transitional phase. Nothing stays
there: either the transition is approved and the spec goes to REC, or
it's not and it reverts to a lower maturity. The things that happen
during PR are useful, but they don't really depend on there being an
explicit phase with an explicit publication, so we can simplify things
by folding all of this into the transition from CR to REC, without a
distinct phase.

Part of https://github.com/w3c/w3process/issues/861

Co-authored-by: Ted Thibodeau Jr <tthibodeau@openlinksw.com>
Co-authored-by: fantasai <fantasai.bugs@inkedblade.net>
Co-authored-by: Nigel Megitt <nigel.megitt@bbc.co.uk>