[Editorial] Give context for candidate amendments when revising a REC
This change gives a little bit more context about how to make normative
changes to a REC. Technically, and this is what the existing text talked
about, they are made by folding in candidate amendments. However,
someone just reading that section may not be aware of what candidate
amendments are and how they are made to begin with.

This gives just a little bit of context to help people piece things
together.

This is a small step towards addressing https://github.com/w3c/w3process/issues/700