Consolidate similar parts of REC revision
Revising a REC is a fairly complicated piece of the Process. The 4
subsections that deal with making revisions for class 1 through 4 are
comparatively simple, but they had very strong similarities between
class 1 and 2, and class 3 and 4. Consolidating the text not only makes
the whole thing shorter, it also eliminates subtle differences of
language that could leave people wondering about potential differences
where none were intended or useful.

This is a small part in addressing https://github.com/w3c/w3process/issues/700