Remove note about AC consensus assessment
Assessing the consensus of the AC in AC reviews is overseen by the Team
as a whole, not by the CEO singularly. Assessing the consensus of the AC
outside of an AC Review is not something any of our processes depend on,
so discussing who is responsible for it is unnecessary (and the phrasing
is clumsy anyway).

See #615