Forbid the Team from AB/TAG service
The Team already has a clearly defined role in the process.
Being responsible for appointments to the TAG, for instance,
would make service on the TAG a bit of a conflict.
Another example is the role of the Team in Councils,
for which clearer rules would be clarifying.

I believe that this was a bit of a problem in the distant past,
but during my involvement with the organization,
this has never come up.
This is offered as a clarification,
not as a reaction to a particular situation.

This is part 2 of the recommendations in #921.