Use appeal vote as escalation path for rejecting charter refinement
This takes inspiration from parts of a proposal by @ianbjacobs.

Invoking W3C Councils is a time consuming and expensive endeavour, so we
should look at cheaper ways to address escalation against the decision
not to start charter refinement.

Since charter refinement is unlikely to be successful if not enough
members are interested, dismissing the escalation when it is supported
by fewer than 5 members seems like a cost effective filter.
When there are 5 or more, an appeals vote is a more cost effective way
to decide the question than an Council, and since this is only about the
start of charter refinement, deep discussions, such as those that a
Council would enable, don't seem necessary.