Enable 5 members to push for resuming an abandoned charter refinement
This takes inspiration from parts of a proposal by @ianbjacobs.

The Council is a powerful but expensive way to resolve disagreements.

If a very limited number of members oppose abandoning a
charter-in-progress, the charter would fail AC Review anyway, and
abandoning was the right decision.

If 5 or more Members insist that abandoning the charter refinement was
the wrong decision, that is a strong indication that there is enough
interest and that the decision was premature. The charter may not quite
be ready for AC Review yet, but enough people have hope that it can
become ready that resuming discussions is the appropriate default
reaction. The facilitator remains the arbiter of how much more effort to
spend on the attempt, and of when to send the draft to the AC for
review.

Nevertheless, the Team may judge that despite pronounced member
interest, the situation is inextricable, and that no charter can be
produced that could be worth sending to the AC for review. In that case,
the Team can refer the case to the Council, hoping to gain approval for
the decision to abandon.