Correct council text pertaining to multiple objections / decisions
This is a follow up to https://github.com/w3c/process/pull/1045

The earlier commits established terminology to discuss Councils in terms
of the decisions they may or may not overturn, and made the Council
section of the Process clearer. However, in the process of doing that,
it broke some of the ways other sections refered to this.

This closes the loop, and makes everything coherent.

Co-authored-by: fantasai <fantasai.bugs@inkedblade.net>